Change limits on training and test image shape
Use the same image shape for training and testing. Using one size for
training and another size for testing would increase difficulty of  the
problem. We don't have this requirement in our problem, so we use the
same size for training and testing to simplify the problem.

Allows the use of image shapes of any aspect ratio. Original question
limits the input image to a square shape, which is not suitable for most
questionnaire images. We need a more flexible input image shape to
meet our requirements.